+++ mobile/android/android-components/components/service/pocket/src/main/java/mozilla/components/service/pocket/PocketStory.kt
@@ -42,6 +42,7 @@ sealed class PocketStory {
     /**
      * A Pocket sponsored story.
      *
+     * @property id Unique id of this story.
      * @property title The title of the story.
      * @property url 3rd party url containing the original story.
      * @property imageUrl A url to a still image representing the story.
@@ -49,13 +50,19 @@ sealed class PocketStory {
      * with a specific resolution and the CENTER_CROP ScaleType.
      * @property sponsor 3rd party sponsor of this story, e.g. "NextAdvisor".
      * @property shim Unique identifiers for when the user interacts with this story.
+     * @property priority Priority level in deciding which stories to be shown first.
+     * A lowest number means a higher priority.
+     * @property caps Story caps indented to control the maximum number of times the story should be shown.
      */
     data class PocketSponsoredStory(
+        val id: Int,
         override val title: String,
         override val url: String,
         val imageUrl: String,
         val sponsor: String,
         val shim: PocketSponsoredStoryShim,
+        val priority: Int,
+        val caps: PocketSponsoredStoryCaps,
     ) : PocketStory()
 
     /**
@@ -67,5 +74,26 @@ sealed class PocketStory {
     data class PocketSponsoredStoryShim(
         val click: String,
         val impression: String,
+    )
+
+    /**
+     * Sponsored story caps indented to control the maximum number of times the story should be shown.
+     *
+     * @property currentImpressions List of all recorded impression of a sponsored Pocket story
+     * expressed in seconds from Epoch (as the result of `System.currentTimeMillis / 1000`).
+     * @property lifetimeCount Lifetime maximum number of times this story should be shown.
+     * This is independent from the count based on [flightCount] and [flightPeriod] and must never be reset.
+     * @property flightCount Maximum number of times this story should be shown in [flightPeriod].
+     * @property flightPeriod Period expressed as a number of seconds in which this story should be shown
+     * for at most [flightCount] times.
+     * Any time the period comes to an end the [flightCount] count should be restarted.
+     * Even if based on [flightCount] and [flightCount] this story can still be shown a couple more times
+     * if [lifetimeCount] was met then the story should not be shown anymore.
+     */
+    data class PocketSponsoredStoryCaps(
+        val currentImpressions: List<Long> = emptyList(),
+        val lifetimeCount: Int,
+        val flightCount: Int,
+        val flightPeriod: Int,
     )
 }