+++ mobile/android/android-components/components/service/pocket/src/main/java/mozilla/components/service/pocket/ext/Mappers.kt
@@ -7,6 +7,7 @@ package mozilla.components.service.pocket.ext
 import androidx.annotation.VisibleForTesting
 import mozilla.components.service.pocket.PocketStory.PocketRecommendedStory
 import mozilla.components.service.pocket.PocketStory.PocketSponsoredStory
+import mozilla.components.service.pocket.PocketStory.PocketSponsoredStoryCaps
 import mozilla.components.service.pocket.PocketStory.PocketSponsoredStoryShim
 import mozilla.components.service.pocket.spocs.api.ApiSpoc
 import mozilla.components.service.pocket.spocs.db.SpocEntity
@@ -53,25 +54,39 @@ internal fun PocketRecommendedStory.toPartialTimeShownUpdate(): PocketLocalStory
  */
 internal fun ApiSpoc.toLocalSpoc(): SpocEntity =
     SpocEntity(
+        id = flightId,
         url = url,
         title = title,
         imageUrl = imageSrc,
         sponsor = sponsor,
         clickShim = shim.click,
-        impressionShim = shim.impression
+        impressionShim = shim.impression,
+        priority = priority,
+        lifetimeCapCount = caps.lifetimeCount,
+        flightCapCount = caps.flightCount,
+        flightCapPeriod = caps.flightPeriod,
     )
 
 /**
  * Map Room entities to the object type that we expose to service clients.
  */
-internal fun SpocEntity.toPocketSponsoredStory(): PocketSponsoredStory =
-    PocketSponsoredStory(
-        title = title,
-        url = url,
-        imageUrl = imageUrl,
-        sponsor = sponsor,
-        shim = PocketSponsoredStoryShim(
-            click = clickShim,
-            impression = impressionShim
-        )
-    )
+internal fun SpocEntity.toPocketSponsoredStory(
+    impressions: List<Long> = emptyList()
+) = PocketSponsoredStory(
+    id = id,
+    title = title,
+    url = url,
+    imageUrl = imageUrl,
+    sponsor = sponsor,
+    shim = PocketSponsoredStoryShim(
+        click = clickShim,
+        impression = impressionShim
+    ),
+    priority = priority,
+    caps = PocketSponsoredStoryCaps(
+        currentImpressions = impressions,
+        lifetimeCount = lifetimeCapCount,
+        flightCount = flightCapCount,
+        flightPeriod = flightCapPeriod,
+    ),
+)