+++ mobile/android/android-components/components/service/pocket/src/test/java/mozilla/components/service/pocket/ext/MappersKtTest.kt
@@ -8,6 +8,7 @@ import mozilla.components.service.pocket.helpers.PocketTestResources
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotEquals
 import org.junit.Assert.assertSame
+import org.junit.Assert.assertTrue
 import org.junit.Test
 import kotlin.reflect.full.memberProperties
 
@@ -78,12 +79,17 @@ class MappersKtTest {
 
         val result = apiStory.toLocalSpoc()
 
+        assertEquals(apiStory.flightId, result.id)
         assertSame(apiStory.title, result.title)
         assertSame(apiStory.url, result.url)
         assertSame(apiStory.imageSrc, result.imageUrl)
         assertSame(apiStory.sponsor, result.sponsor)
         assertSame(apiStory.shim.click, result.clickShim)
         assertSame(apiStory.shim.impression, result.impressionShim)
+        assertEquals(apiStory.priority, result.priority)
+        assertEquals(apiStory.caps.lifetimeCount, result.lifetimeCapCount)
+        assertEquals(apiStory.caps.flightCount, result.flightCapCount)
+        assertEquals(apiStory.caps.flightPeriod, result.flightCapPeriod)
     }
 
     @Test
@@ -92,11 +98,17 @@ class MappersKtTest {
 
         val result = localStory.toPocketSponsoredStory()
 
+        assertEquals(localStory.id, result.id)
         assertSame(localStory.title, result.title)
         assertSame(localStory.url, result.url)
         assertSame(localStory.imageUrl, result.imageUrl)
         assertSame(localStory.sponsor, result.sponsor)
         assertSame(localStory.clickShim, result.shim.click)
         assertSame(localStory.impressionShim, result.shim.impression)
+        assertEquals(localStory.priority, result.priority)
+        assertEquals(localStory.lifetimeCapCount, result.caps.lifetimeCount)
+        assertEquals(localStory.flightCapCount, result.caps.flightCount)
+        assertEquals(localStory.flightCapPeriod, result.caps.flightPeriod)
+        assertTrue(result.caps.currentImpressions.isEmpty())
     }
 }